+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -1,2 +1,3 @@
 export 'instagram.dart';
 export 'telegram.dart';
+export 'whatsapp.dart';