+++ lib/src/core/deeplink_x.dart
@@ -88,14 +88,14 @@ class DeeplinkX {
   /// Returns a [Future<bool>] that completes with:
   /// * `true` if the native URI can be launched
   /// * `false` if the native URI cannot be launched
-  /// 
+  ///
   /// Note: apps schemes must be registered based on the platform.
   /// Check Platform-Specific Configurations in apps documents for more details.
   ///
   /// Example:
   /// ```dart
   /// final deeplinkX = DeeplinkX();
-  /// 
+  ///
   /// // Check if native Instagram URI can be launched
   /// final canLaunchNative = await deeplinkX.canLaunchNativeDeeplink(
   ///   Instagram.openProfile('username')